django-pjt/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
axios==0.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-appconf==1.0.5
django-cors-headers==4.6.0
django-environ==0.11.2
django-imagekit==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jiter==0.7.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.55.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pilkit==3.0
Pillow==10.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Wikipedia-API==0.6.0
zipp==3.19.2
